First C++11 version
The previous version of the code used Qt for portable threading
and synchronization primitives. Since the release of C++11, the
threading support library was added very slowly by compiler vendors
but now it really looks like it is possible to write portable C++11
multithreading code.

This release includes a C++ version of the LMAX disruptor pattern.
It is based on template, but now I actually think this was a bad
decision and I plan to revert it back to use simple composition.